Added a program that prints the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line.

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,11 +9,11 @@
 int main(void)
 {
 	int i;
-	unsigned long long num1 = 1, num2 = 2;
+	unsigned long num1 = 1, num2 = 2;
 
 	for (i = 0; i < 50; i += 2)
 	{
-		printf("%llu, %llu\n", num1, num2);
+		printf("%lu, %lu\n", num1, num2);
 
 		num1 = num1 + num2;
 		num2 = num1 + num2;
